Move the Signature type closer to the BLS routines
Create a new `Signature` type in the BLS package.
Modify `BLSSignature` to inherit from it.
Use the new type when calling `bls_verify()`.
